backend/ai-assistant-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
httpx